Move source file
